# internal/openapi/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Quantyra GeoIDX API",
    "description": "OpenAPI document for the **idx-api** Go service (Fiber): PAT auth, versioned Bridge/Spark MLS proxy (`/api/v1/*`), GIS parcel GeoJSON, hybrid search, comps engine, replication stats, and listing photos (`/images/*`). This collection covers JSON/API endpoints only and intentionally excludes HTML/session dashboard pages (`/login`, `/dashboard/*`, `/invite/*`) plus static assets (`/static/*`). Routes are registered in `internal/api/routes.go` and `internal/handler/dashboard/handler.go`.",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "https://idx.quantyralabs.cc",
      "description": "Production (`IDX_PLATFORM_URL` / `IDX_API_PUBLIC_URL` — API and dashboard on the platform host)"
    },
    {
      "url": "http://localhost:8000",
      "description": "Local Go API (`make run-api` or `go run ./cmd/api`, default port 8000)"
    }
  ],
...
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "description": "Bearer PAT on routes guarded by `DomainToken` middleware (`/api/v1/*`, `/images/*`). Abilities: **`idx:full`** (issued by `POST /api/auth/token`) or **`idx:access`**. Always pair with `domainSlugAuth` (`X-Domain-Slug` header) or the `?domain=` query parameter."
      },
      "domainSlugAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "X-Domain-Slug",
        "description": "TXT-verified domain slug required alongside `bearerAuth` for all DomainToken-protected routes (`/api/v1/*`, `/images/*`). Must match a verified domain on the token owner's account. The `?domain=` query parameter is an equivalent alternative accepted by the same middleware. Not required for `/api/auth/user` (token-only) or admin session endpoints."
      }
    },
    "schemas": {
...
      "ListingsResponse": {
        "type": "object",
        "description": "Bridge Data Output web API envelope for GET /api/v1/listings (not OData `value`). RESO OData routes use a separate shape.",
        "properties": {
          "pricing": {
            "type": "object",
            "description": "Global quote matrix used for listing-level conversion output (when include_pricing=1).",
            "properties": {
              "status": { "type": "string", "example": "ok" },
              "as_of": { "type": "string", "format": "date-time", "nullable": true },
              "quotes": {
                "type": "object",
                "additionalProperties": {
                  "type": "object",
                  "additionalProperties": {
                    "type": "number"
                  }
                }
              }
            }
          },
          "bundle": {
            "type": "array",
            "description": "Listing objects from Bridge web API.",
            "items": {
              "type": "object",
              "properties": {
                "ListPrice": { "type": "number", "nullable": true },
                "pricing_converted": {
                  "type": "object",
                  "properties": {
                    "base_currency": { "type": "string", "example": "usd" },
                    "fiat": {
                      "type": "object",
                      "additionalProperties": { "type": "number" }
                    },
                    "digital_assets": {
                      "type": "object",
                      "additionalProperties": { "type": "number" }
                    }
                  }
                }
              },
              "additionalProperties": true
            }
          },
          "status": { "type": "integer", "example": 200 },
          "success": { "type": "boolean", "example": true },
          "total": { "type": "integer", "example": 4708105, "description": "Total listings matching the upstream query." }
        }
      },
...
    }
  },
  "paths": {
...
    "/api/v1/listings": {
      "get": {
        "tags": ["MLS Data - Listings"],
        "summary": "Get Listings",
        "description": "Get MLS listings for the resolved MLS feed (Bridge web collection when `filters` is set; merged Active/Pending Bridge RESO slice when `filters` is absent). Responses include top-level `pricing` and per-row `pricing_converted` when `ListPrice` is present and `include_pricing=1` is passed. **Collection cache:** 15-minute TTL applies per domain when **`filters` is absent**; any `filters` query parameter bypasses the cache and forwards to the upstream MLS. Standard Bridge/Spark query parameters are passed through. Responds with `X-IDX-Cache: HIT|MISS`.",
        "security": [{ "bearerAuth": [], "domainSlugAuth": [] }],
        "parameters": [
          {
            "name": "X-Domain-Slug",
            "in": "header",
            "required": false,
            "schema": { "type": "string", "example": "example-com" },
            "description": "Domain slug (alternative to `?domain=` query param). Must match a verified domain on the token owner's account."
          },
          {
            "name": "domain",
            "in": "query",
            "required": false,
            "schema": { "type": "string", "example": "example-com" },
            "description": "Domain slug (alternative to `X-Domain-Slug` header)."
          },
          {
            "name": "dataset",
            "in": "query",
            "required": false,
            "schema": { "type": "string", "enum": ["stellar", "beaches"], "example": "stellar" },
            "description": "MLS dataset override. Must be enabled for the requesting domain."
          },
          {
            "name": "include_pricing",
            "in": "query",
            "required": false,
            "schema": { "type": "string", "enum": ["1"] },
            "description": "When `1`, injects top-level `pricing` and per-listing `pricing_converted` from the crypto pricing reader."
          },
          {
            "name": "filters",
            "in": "query",
            "required": false,
            "schema": { "type": "string" },
            "description": "Any non-empty value bypasses the cache and proxies directly to the upstream MLS web endpoint."
          }
        ],
        "responses": {
          "200": {
            "description": "Listings retrieved successfully",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ListingsResponse" }
              }
            }
          },
          "401": { "description": "Unauthenticated" },
          "403": { "description": "Forbidden — domain/token mismatch or dataset not allowed" },
          "502": { "description": "Upstream MLS error" }
        }
      }
    },
...
    "/api/v1/agents": {
      "get": {
        "tags": ["MLS Data - Agents & Offices"],
        "summary": "Get Agents",
        "description": "Proxy to Bridge/Spark web agents collection. Standard query parameters (`limit`, OData `$filter`, etc.) are passed through. Auth: `Authorization: Bearer <token>` + `X-Domain-Slug` or `?domain=`.",
        "security": [{ "bearerAuth": [], "domainSlugAuth": [] }],
        "parameters": [
          { "name": "limit", "in": "query", "schema": { "type": "integer", "default": 10 } },
          { "name": "dataset", "in": "query", "required": false, "schema": { "type": "string", "enum": ["stellar", "beaches"] } }
        ],
        "responses": {
          "200": { "description": "Agent list" },
          "401": { "description": "Unauthenticated" },
          "502": { "description": "Upstream MLS error" }
        }
      }
    },
...
    "/api/v1/offices": {
      "get": {
        "tags": ["MLS Data - Agents & Offices"],
        "summary": "Get Offices",
        "description": "Proxy to Bridge/Spark web offices collection. Auth: `Authorization: Bearer <token>` + `X-Domain-Slug` or `?domain=`.",
        "security": [{ "bearerAuth": [], "domainSlugAuth": [] }],
        "parameters": [
          { "name": "limit", "in": "query", "schema": { "type": "integer", "default": 10 } },
          { "name": "dataset", "in": "query", "required": false, "schema": { "type": "string", "enum": ["stellar", "beaches"] } }
        ],
        "responses": {
          "200": { "description": "Office list" },
          "401": { "description": "Unauthenticated" },
          "502": { "description": "Upstream MLS error" }
        }
      }
    },
...
    "/api/v1/openhouses": {
      "get": {
        "tags": ["MLS Data - Open Houses"],
        "summary": "Get Open Houses",
        "description": "Proxy to Bridge/Spark web open houses collection. Auth: `Authorization: Bearer <token>` + `X-Domain-Slug` or `?domain=`.",
        "security": [{ "bearerAuth": [], "domainSlugAuth": [] }],
        "parameters": [
          { "name": "limit", "in": "query", "schema": { "type": "integer", "default": 10 } },
          { "name": "dataset", "in": "query", "required": false, "schema": { "type": "string", "enum": ["stellar", "beaches"] } }
        ],
        "responses": {
          "200": { "description": "Open house list" },
          "401": { "description": "Unauthenticated" },
          "502": { "description": "Upstream MLS error" }
        }
      }
    },
...
    "/api/v1/pub/parcels": {
      "get": {
        "tags": ["MLS Data - Public Records"],
        "summary": "Get Public Parcels",
        "description": "Proxy to the Bridge Public Records `/pub/parcels` collection. Returns parcel records with property identifiers, county data, and assessment linkage. Auth: `Authorization: Bearer <token>` + `X-Domain-Slug` or `?domain=`.",
        "security": [{ "bearerAuth": [], "domainSlugAuth": [] }],
        "responses": {
          "200": { "description": "Parcel list" },
          "401": { "description": "Unauthenticated" },
          "502": { "description": "Upstream Bridge error" }
        }
      }
    },
...
    "/api/v1/pub/assessments": {
      "get": {
        "tags": ["MLS Data - Public Records"],
        "summary": "Get Public Assessments",
        "description": "Proxy to Bridge public records assessment collection (all parcels). Standard OData parameters are passed through.",
        "security": [{ "bearerAuth": [], "domainSlugAuth": [] }],
        "responses": {
          "200": { "description": "Assessment list" },
          "401": { "description": "Unauthenticated" },
          "502": { "description": "Upstream Bridge error" }
        }
      }
    },
    "/api/v1/pub/transactions": {
      "get": {
        "tags": ["MLS Data - Public Records"],
        "summary": "Get Public Transactions",
        "description": "Proxy to Bridge public records transaction collection (all parcels). Standard OData parameters are passed through.",
        "security": [{ "bearerAuth": [], "domainSlugAuth": [] }],
        "responses": {
          "200": { "description": "Transaction list" },
          "401": { "description": "Unauthenticated" },
          "502": { "description": "Upstream Bridge error" }
        }
      }
    },
...
  }
}